refactor(chat): extract subagent panel controller
